Update inundate job tests and documentation for depths
This commit fixes the inundate job tests to work with the new pointer
parquet file format for catchment data. It also revises the tests so
that inunudated extents with greater than 99% similarity to the test
data pass to account for minor rounding errors in the depth extents.

In terms of documentation interfaces.md was edited to reflect new
catchment data file removing some of the arguments that inundate.py
previously accepted. A doc string in inundate.py was also edited.